requirements.txt
Kivy==2.3.0
opencv-python==4.10.0.82
python-dotenv==1.0.1
